table mistake fixed

+++ sql/copies_table_creation.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS copies CASCADE;
 CREATE TABLE copies (
                         isbn text NOT NULL,
-                        copy_id bigint PRIMARY KEY,
+                        copy_id SERIAL PRIMARY KEY,
                         availability text NOT NULL,
                         FOREIGN KEY (isbn) references books(isbn)
 );
