+++ sql/history_table_creation.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS history CASCADE;
 CREATE TABLE history (
                          isbn text NOT NULL,
-                         copy_id bigint NOT NULL,
+                         copy_id integer NOT NULL,
                          reader_id integer NOT NULL,
                          issue_date date NOT NULL,
                          return_date date NOT NULL,
